% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LinReg.R
\name{LinReg}
\alias{LinReg}
\title{A simple multiple linear regression function (OLS)}
\usage{
LinReg(dv, iv, data, bootstrap, Number_Bootstrapps, outlier_controll, plot)
}
\arguments{
\item{dv}{dependent vaiable}

\item{iv}{a string vector with the names of the independent variables, separated by commas}

\item{data}{a data frame containing the variables}

\item{bootstrap}{set bootstap to TRUE or FALSE, if FALSE Number of bootstraps are ignored}

\item{outlier_controll}{set outlier_controll to TRUE or FALSE, to use cooks distance to exlude outliers, if bootstrap==TRUE, outlier_controll must be FALSE}

\item{set}{plot to TRUE to create simple scatterplots of correlation between variables}
}
\value{
the results of linear regression, plots and all requirements plus an conclusion about the violations
}
\description{
This function allows you to perform a linear regression plus requirements
}
\examples{
LinReg('Var_1',c('Var_2','Var_3','Var_4'),test, bootstrap = TRUE, Number_Bootstrapps=1000, outlier_controll = FALSE, plot=TRUE)
}
\keyword{Linear}
\keyword{Regression}
